Implemented a warning for rate names which are defined but not used. To do this, it made sense to make TopRate an expression, since we have to visit all of the expressions in a model. Now TopRate can only occur (because of the parser) at the top level of an expression in a prefix node, it cannot occur generally within an expression. Even still, I felt that this was a bit cleaner than checking for that. This way, if we ever do allow TopRate to be part of an expression (for example even as simply a constant definition) then we won't have to make TopRate an expression because it already will be.
As for our representation of warnings and errors, they are currently as a
list. This is a little bogus since we probably do not wish to make the same
warning more than once, for example we may have several uses of an
undefined rate name but we likely only wish to report that once. Although
there are other ways around that. This made sense because we test using
equality on the warning representation, which is a tad cludgy but means
that we are not testing on the strings thus produced (though we could
relatively easily do that).